requirements.txt
aiofiles==24.1.0
async-timeout==5.0.1
asyncio-redis==0.16.0
asyncpg==0.30.0
greenlet==3.1.1
hiredis==3.1.0
html5tagger==1.3.0
httptools==0.6.4
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
PyYAML==6.0.2
sanic==24.12.0
sanic-ext==24.12.0
sanic-routing==23.12.0
sanic-session==0.8.0
setuptools==76.0.0
SQLAlchemy==2.0.39
tracerite==1.1.1
typing_extensions==4.12.2
ujson==5.10.0
websockets==15.0.1

requests~=2.32.3
redis~=5.2.1